+++ src/Grisette/Experimental/Qualified/ParallelUnionDo.hs
@@ -6,13 +6,13 @@
 -- Maintainer  :   siruilu@cs.washington.edu
 -- Stability   :   Experimental
 -- Portability :   GHC only
-module Grisette.Qualified.ParallelUnionDo ((>>=), (>>)) where
+module Grisette.Experimental.Qualified.ParallelUnionDo ((>>=), (>>)) where
 
 import Control.Parallel.Strategies (NFData)
-import Grisette.Core.Control.Monad.Class.MonadParallelUnion
+import Grisette.Core.Data.Class.Mergeable (Mergeable)
+import Grisette.Experimental.MonadParallelUnion
   ( MonadParallelUnion (parBindUnion),
   )
-import Grisette.Core.Data.Class.Mergeable (Mergeable)
 import Prelude (const, ($))
 
 (>>=) :: (MonadParallelUnion m, Mergeable b, NFData b) => m a -> (a -> m b) -> m b
